27 Apr 2016

+++ Password_Backup/page_2.cpp
@@ -6,12 +6,16 @@
 #include <QTextStream>
 #include <QDebug>
 #include <QList>
+#include <QDesktopWidget>
+#include <QRect>
 
 Page_2::Page_2(QWidget *parent) :
     QDialog(parent),
     ui(new Ui::Page_2)
 {
     ui->setupUi(this);
+    const QRect screen = QApplication::desktop()->screenGeometry();
+    this->move(screen.center()- this->rect().center());
 
   // QStandardItemModel *model = new QStandardItemModel();
    model->setRowCount(0);
@@ -37,11 +41,11 @@ Page_2::Page_2(QWidget *parent) :
        list.append(user);
        list.append(passcode);
       // qDebug() << item1->text();
-       model->setHeaderData(0,Qt::Horizontal,QObject::tr("Account"));
-       model->setHeaderData(1,Qt::Horizontal,QObject::tr("Username"));
-       model->setHeaderData(2,Qt::Horizontal,QObject::tr("Password"));
        model->appendRow(list);
    }
+   model->setHeaderData(0,Qt::Horizontal,QObject::tr("Account"));
+   model->setHeaderData(1,Qt::Horizontal,QObject::tr("Username"));
+   model->setHeaderData(2,Qt::Horizontal,QObject::tr("Password"));
    file.close();
 }
 
@@ -52,8 +56,20 @@ Page_2::~Page_2()
 
 void Page_2::on_pushButton_clear_pws_clicked()
 {
-    QString filename = "passwords.txt";
-    QFile file(filename);
-    file.remove();
-    model->clear();
+    QMessageBox::StandardButton reply;
+    reply = QMessageBox::question(this,"","Click yes to delete all passwords",QMessageBox::Yes | QMessageBox::No);
+    if(reply == QMessageBox::Yes)
+    {
+        QString filename = "passwords.txt";
+        QFile file(filename);
+        file.remove();
+        model->clear();
+    }
+}
+
+void Page_2::on_pushButton_last_back_clicked()
+{
+    bool goBack = true;
+    emit send_button_status2(goBack);
+    this->deleteLater();
 }
